pyproject.toml
[project]
name = "gridrl"
version = "0.1.1"
authors = [
    {name = "Sky", email = "gridrlgym@gmail.com"}
]
keywords = ["GridWorld", "Exploration", "AI", "RL"]
description = "Customizable engine for minimalist 2D grid-based games, oriented towards Reinforcement Learning."
readme = "README.md"
license = {file = "LICENSE.md"}
classifiers = [
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"

dependencies = [
    "numpy",
    "pillow",
    "gymnasium>=0.29",
    "matplotlib",
    "pandas",
]

[project.optional-dependencies]
all = []

[project.urls]
Homepage = "https://github.com/GridRL/GridRL"
#Documentation = "https://github.com/GridRL/GridRL/docs/index.html"
Repository = "https://github.com/GridRL/gridrl.git"

[project.scripts]
gridrl = "gridrl.cli:main_cli"

[build-system]
requires = [
    "setuptools>=61.0.0",
    "numpy",
    "cython>=3.0.6; platform_python_implementation == 'CPython'",
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
zip-safe = false
packages = ["gridrl","gridrl.abstract_games","gridrl.agents","gridrl.envs","gridrl.games",
    "gridrl.games.exploration_world1","gridrl.games.creatures_world1"
]

[tool.setuptools.package-data]
"*" = [
    "gridrl/games/**/assets/**/*.png",
    "gridrl/games/**/data/*.json",
    "gridrl/games/**/data/*.json.gz",
]

[tool.ruff.lint.per-file-ignores]
"gridrl/core_game.py" = []
